# laravel11/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SyncBridge API",
        "description": "Production-grade API documentation for SyncBridge integration endpoints.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8080",
            "description": "Primary API server"
        }
    ],
    "paths": {
...
        "/api/v1/sync": {
            "post": {
                "tags": [
                    "Sync"
                ],
                "summary": "Run sync command",
                "description": "Executes a supported sync command and returns run metadata.",
                "operationId": "syncRun",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SyncCommandRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sync command executed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiSuccessBase"
                                        },
                                        {
                                            "properties": {
                                                "run_id": {
                                                    "type": "integer",
                                                    "example": 1001,
                                                    "nullable": true
                                                },
                                                "result": {
                                                    "type": "object",
                                                    "additionalProperties": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Sync execution failure",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "apiTokenHeader": []
                    }
                ]
            }
        },
        "/api/v1/sync/products": {
            "post": {
                "tags": [
                    "Sync"
                ],
                "summary": "Run product sync for specific product IDs",
                "description": "Runs products sync for the provided PrestaShop product IDs. Defaults to incremental mode.",
                "operationId": "syncRunProductsByIds",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductSyncRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product sync command executed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiSuccessBase"
                                        },
                                        {
                                            "properties": {
                                                "run_id": {
                                                    "type": "string",
                                                    "example": "f0e1d2c3b4a59687",
                                                    "nullable": true
                                                },
                                                "result": {
                                                    "type": "object",
                                                    "additionalProperties": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Sync execution failure",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    },
                    {
                        "apiTokenHeader": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ApiSuccessBase": {
                "required": [
                    "ok",
                    "message"
                ],
                "properties": {
                    "ok": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "ok"
                    }
                },
                "type": "object"
            },
            "ApiError": {
                "required": [
                    "ok",
                    "error",
                    "errors"
                ],
                "properties": {
                    "ok": {
                        "type": "boolean",
                        "example": false
                    },
                    "error": {
                        "type": "string",
                        "example": "Invalid payload"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "schema": "OpenApiSpec",
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "The field is required."
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "SyncCommandRequest": {
                "required": [
                    "command"
                ],
                "properties": {
                    "command": {
                        "type": "string",
                        "enum": [
                            "products",
                            "products:inc",
                            "stock",
                            "orders",
                            "order-status",
                            "all",
                            "retry",
                            "status"
                        ],
                        "example": "products"
                    },
                    "since": {
                        "type": "string",
                        "example": "2026-05-01T00:00:00Z",
                        "nullable": true
                    },
                    "ps_product_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": 123
                        }
                    }
                },
                "type": "object"
            },
            "ProductSyncRequest": {
                "required": [
                    "ps_product_ids"
                ],
                "properties": {
                    "ps_product_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 92
                        }
                    },
                    "sync_command": {
                        "type": "string",
                        "enum": [
                            "products",
                            "products:inc"
                        ],
                        "example": "products:inc"
                    },
                    "since": {
                        "type": "string",
                        "example": "2026-05-01 00:00:00",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Paste SYNCBRIDGE_API_TOKEN as Bearer token.",
                "bearerFormat": "Token",
                "scheme": "bearer"
            },
            "apiTokenHeader": {
                "type": "apiKey",
                "description": "Alternative auth header for sync-protected endpoints.",
                "name": "X-Api-Token",
                "in": "header"
            }
        }
    },
...
}